+++ test/test_unary_ufuncs.py
@@ -13,7 +13,7 @@ import unittest
 from torch._six import inf, nan
 from torch.testing._internal.common_utils import (
     TestCase, run_tests, torch_to_numpy_dtype_dict, numpy_to_torch_dtype_dict,
-    suppress_warnings, TEST_SCIPY, slowTest, skipIfNoSciPy, IS_WINDOWS)
+    suppress_warnings, TEST_SCIPY, slowTest, skipIfNoSciPy, IS_WINDOWS, gradcheck)
 from torch.testing._internal.common_methods_invocations import (
     unary_ufuncs, _NOTHING)
 from torch.testing._internal.common_device_type import (
@@ -1017,6 +1017,21 @@ class TestUnaryUfuncs(TestCase):
         self.assertEqual(torch.nn.functional.silu(
             input_noncontig, inplace=True), expected_output_noncontig,
             atol=atol, rtol=rtol)
+
+    # It is not obvious how to merge this into OpInfo becuase these inputs
+    # succeed for gradcheck but are expected to fail for gradgradcheck
+    @dtypes(torch.double)
+    def test_sinc(self, device, dtype):
+        # The derivative of sinc(x) at x=0 has to be special cased.
+        # A naive computation will result in 0/0 -> NaN.
+        # We also need to be careful when we are very close to 0, as the
+        # derivative's denominator is squared, and there are some floats
+        # that are positive and whose squares are zero.
+        a = torch.tensor([0.0, torch.finfo(torch.double).tiny, 1.0],
+                         dtype=dtype,
+                         requires_grad=True,
+                         device=device)
+        gradcheck(torch.sinc, a)
 
     @skipIfNoSciPy
     @dtypes(torch.float, torch.double)
